+++ core/cmake/BareosFindAllLibraries.cmake
@@ -1,6 +1,6 @@
 #   BAREOS® - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2017-2022 Bareos GmbH & Co. KG
+#   Copyright (C) 2017-2023 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -104,8 +104,6 @@ endif()
 if(NOT ${CMAKE_SYSTEM_NAME} MATCHES "Windows")
   include(FindPostgreSQL)
 endif()
-
-include(CMakeUserFindMySQL)
 
 if(${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
   set(OPENSSL_USE_STATIC_LIBS 1)