+++ core/src/cats/bvfs.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2009-2010 Free Software Foundation Europe e.V.
    Copyright (C) 2016-2016 Planets Communications B.V.
-   Copyright (C) 2016-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2016-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -26,7 +26,7 @@
  */
 #include "include/bareos.h"
 
-#if HAVE_SQLITE3 || HAVE_MYSQL || HAVE_POSTGRESQL || HAVE_INGRES || HAVE_DBI
+#if HAVE_POSTGRESQL
 
 #  include "cats/cats.h"
 #  include "cats/sql.h"
@@ -72,18 +72,14 @@ void BareosDb::BuildPathHierarchy(JobControlRecord* jcr,
 
   Dmsg1(dbglevel, "BuildPathHierarchy(%s)\n", new_path);
 
-  /*
-   * Does the ppathid exist for this? use a memory cache ...
+  /* Does the ppathid exist for this? use a memory cache ...
    * In order to avoid the full loop, we consider that if a dir is already in
    * the PathHierarchy table, then there is no need to calculate all the
-   * hierarchy
-   */
+   * hierarchy */
   while (new_path && *new_path) {
     if (ppathid_cache.lookup(pathid)) {
-      /*
-       * It's already in the cache.  We can leave, no time to waste here,
-       * all the parent dirs have already been done
-       */
+      /* It's already in the cache.  We can leave, no time to waste here,
+       * all the parent dirs have already been done */
       goto bail_out;
     } else {
       Mmsg(cmd, "SELECT PPathId FROM PathHierarchy WHERE PathId = %llu",
@@ -165,12 +161,10 @@ bool BareosDb::UpdatePathHierarchyCache(JobControlRecord* jcr,
   Mmsg(cmd, "UPDATE Job SET HasCache=-1 WHERE JobId=%s", jobid);
   UPDATE_DB(jcr, cmd);
 
-  /*
-   * need to COMMIT here to ensure that other concurrent .bvfs_update runs
+  /* need to COMMIT here to ensure that other concurrent .bvfs_update runs
    * see the current HasCache value. A new transaction must only be started
    * after having finished PathHierarchy processing, otherwise prevention
-   * from duplicate key violations in BuildPathHierarchy() will not work.
-   */
+   * from duplicate key violations in BuildPathHierarchy() will not work. */
   EndTransaction(jcr);
 
   /* Inserting path records for JobId */
@@ -189,12 +183,10 @@ bool BareosDb::UpdatePathHierarchyCache(JobControlRecord* jcr,
     goto bail_out;
   }
 
-  /*
-   * Now we have to do the directory recursion stuff to determine missing
+  /* Now we have to do the directory recursion stuff to determine missing
    * visibility.
    * We try to avoid recursion, to be as fast as possible.
-   * We also only work on not already hierarchised directories ...
-   */
+   * We also only work on not already hierarchised directories ... */
   Mmsg(cmd,
        "SELECT PathVisibility.PathId, Path "
        "FROM PathVisibility "
@@ -440,17 +432,13 @@ char* bvfs_parent_dir(char* path)
     p += len;
     while (p > path && !IsPathSeparator(*p)) { p--; }
     if (IsPathSeparator(*p) and (len >= 1)) {
-      /*
-       * Terminate the string after the "/".
+      /* Terminate the string after the "/".
        * Do this instead of overwritting the "/"
-       * to keep the root directory "/" as a separate path.
-       */
+       * to keep the root directory "/" as a separate path. */
       p[1] = '\0';
     } else {
-      /*
-       * path did not start with a "/".
-       * This can be the case for plugin results.
-       */
+      /* path did not start with a "/".
+       * This can be the case for plugin results. */
       p[0] = '\0';
     }
   }
@@ -684,23 +672,7 @@ static bool CheckTemp(char* output_table)
 
 void Bvfs::clear_cache()
 {
-  /*
-   * FIXME:
-   * can't use predefined query,
-   * as MySQL queries do only support single SQL statements,
-   * not multiple.
-   */
-  // db->SqlQuery(BareosDb::SQL_QUERY::bvfs_clear_cache_0);
-  db->StartTransaction(jcr);
-  db->SqlQuery("UPDATE Job SET HasCache=0");
-  if (db->GetTypeIndex() == SQL_TYPE_SQLITE3) {
-    db->SqlQuery("DELETE FROM PathHierarchy;");
-    db->SqlQuery("DELETE FROM PathVisibility;");
-  } else {
-    db->SqlQuery("TRUNCATE PathHierarchy");
-    db->SqlQuery("TRUNCATE PathVisibility");
-  }
-  db->EndTransaction(jcr);
+  db->SqlQuery(BareosDb::SQL_QUERY::bvfs_clear_cache_0);
 }
 
 bool Bvfs::DropRestoreList(char* output_table)
@@ -864,17 +836,6 @@ bool Bvfs::compute_restore_list(char* fileid,
     goto bail_out;
   }
 
-  /* MySQL need it */
-  if (db->GetTypeIndex() == SQL_TYPE_MYSQL) {
-    Mmsg(query, "CREATE INDEX idx_%s ON %s (JobId)", output_table,
-         output_table);
-    Dmsg1(dbglevel_sql, "q=%s\n", query.c_str());
-    if (!db->SqlQuery(query.c_str())) {
-      Dmsg0(dbglevel, "Can't execute q\n");
-      goto bail_out;
-    }
-  }
-
   retval = true;
 
 bail_out:
@@ -883,5 +844,4 @@ bail_out:
   return retval;
 }
 
-#endif /* HAVE_SQLITE3 || HAVE_MYSQL || HAVE_POSTGRESQL || HAVE_INGRES || \
-          HAVE_DBI */
+#endif /* HAVE_POSTGRESQL */