+++ core/src/cats/ddl/updates/mysql.2003_2004.sql
@@ -1,4 +1,0 @@
-
-ALTER TABLE FileSet ADD COLUMN FileSetText BLOB NOT NULL;
-
-UPDATE Version SET VersionId = 2004;