+++ core/src/cats/ddl/updates/sqlite3.2003_2004.sql
@@ -1,6 +1,0 @@
-BEGIN;
-
-ALTER TABLE FileSet ADD COLUMN FileSetText TEXT DEFAULT '';
-
-UPDATE Version SET VersionId = 2004;
-COMMIT;