+++ core/src/cats/dml/create_queryfiles.sh
@@ -2,7 +2,7 @@
 
 #   BAREOS® - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2017-2021 Bareos GmbH & Co. KG
+#   Copyright (C) 2017-2023 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -20,7 +20,7 @@
 #   02110-1301, USA.
 
 
-DATABASES="sqlite postgresql mysql"
+DATABASES="postgresql"
 QUERY_NAMES_FILE="../bdb_query_names.inc"
 QUERY_ENUM_FILE="../bdb_query_enum_class.h"