+++ core/src/cats/sql_list.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2009 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -28,7 +28,7 @@
 
 #include "include/bareos.h"
 
-#if HAVE_SQLITE3 || HAVE_MYSQL || HAVE_POSTGRESQL || HAVE_INGRES || HAVE_DBI
+#if HAVE_POSTGRESQL
 
 #  include "cats.h"
 #  include "lib/edit.h"
@@ -227,12 +227,10 @@ void BareosDb::ListMediaRecords(JobControlRecord* jcr,
 
   EscapeString(jcr, esc, mdbr->VolumeName, strlen(mdbr->VolumeName));
 
-  /*
-   * There is one case where ListMediaRecords() is called from SelectMediaDbr()
+  /* There is one case where ListMediaRecords() is called from SelectMediaDbr()
    * with the range argument set to NULL. To avoid problems, we set the range to
    * an empty string if range is set to NULL. Otherwise it would result in
-   * malformed SQL queries.
-   */
+   * malformed SQL queries. */
   if (range == NULL) { range = ""; }
 
   if (count) {
@@ -436,8 +434,7 @@ void BareosDb::ListLogRecords(JobControlRecord* jcr,
   }
 
   if (type != VERT_LIST) {
-    /*
-     * When something else then a vertical list is requested set the list type
+    /* When something else then a vertical list is requested set the list type
      * to RAW_LIST e.g. non formated raw data as that makes the only sense for
      * the logtext output. The logtext already has things like \n etc in it
      * so we should just dump the raw content out for the best visible output.
@@ -473,8 +470,7 @@ void BareosDb::ListJoblogRecords(JobControlRecord* jcr,
   } else {
     FillQuery(SQL_QUERY::list_joblog_2, edit_int64(JobId, ed1), range);
     if (type != VERT_LIST) {
-      /*
-       * When something else then a vertical list is requested set the list type
+      /* When something else then a vertical list is requested set the list type
        * to RAW_LIST e.g. non formated raw data as that makes the only sense for
        * the logtext output. The logtext already has things like \n etc in it
        * so we should just dump the raw content out for the best visible output.
@@ -665,31 +661,17 @@ void BareosDb::ListFilesForJob(JobControlRecord* jcr,
 
   DbLocker _{this};
 
-  if (GetTypeIndex() == SQL_TYPE_MYSQL) {
-    Mmsg(cmd,
-         "SELECT CONCAT(Path.Path,Name) AS Filename "
-         "FROM (SELECT PathId, Name FROM File WHERE JobId=%s "
-         "UNION ALL "
-         "SELECT PathId, Name "
-         "FROM BaseFiles JOIN File "
-         "ON (BaseFiles.FileId = File.FileId) "
-         "WHERE BaseFiles.JobId = %s"
-         ") AS F, Path "
-         "WHERE Path.PathId=F.PathId",
-         edit_int64(jobid, ed1), ed1);
-  } else {
-    Mmsg(cmd,
-         "SELECT Path.Path||Name AS Filename "
-         "FROM (SELECT PathId, Name FROM File WHERE JobId=%s "
-         "UNION ALL "
-         "SELECT PathId, Name "
-         "FROM BaseFiles JOIN File "
-         "ON (BaseFiles.FileId = File.FileId) "
-         "WHERE BaseFiles.JobId = %s"
-         ") AS F, Path "
-         "WHERE Path.PathId=F.PathId",
-         edit_int64(jobid, ed1), ed1);
-  }
+  Mmsg(cmd,
+       "SELECT Path.Path||Name AS Filename "
+       "FROM (SELECT PathId, Name FROM File WHERE JobId=%s "
+       "UNION ALL "
+       "SELECT PathId, Name "
+       "FROM BaseFiles JOIN File "
+       "ON (BaseFiles.FileId = File.FileId) "
+       "WHERE BaseFiles.JobId = %s"
+       ") AS F, Path "
+       "WHERE Path.PathId=F.PathId",
+       edit_int64(jobid, ed1), ed1);
 
   sendit->ArrayStart("filenames");
   if (!BigSqlQuery(cmd, ::ListResult, &lctx)) { return; }
@@ -707,23 +689,13 @@ void BareosDb::ListBaseFilesForJob(JobControlRecord* jcr,
 
   DbLocker _{this};
 
-  if (GetTypeIndex() == SQL_TYPE_MYSQL) {
-    Mmsg(cmd,
-         "SELECT CONCAT(Path.Path,File.Name) AS Filename "
-         "FROM BaseFiles, File, Path "
-         "WHERE BaseFiles.JobId=%s AND BaseFiles.BaseJobId = File.JobId "
-         "AND BaseFiles.FileId = File.FileId "
-         "AND Path.PathId=File.PathId",
-         edit_int64(jobid, ed1));
-  } else {
-    Mmsg(cmd,
-         "SELECT Path.Path||File.Name AS Filename "
-         "FROM BaseFiles, File, Path "
-         "WHERE BaseFiles.JobId=%s AND BaseFiles.BaseJobId = File.JobId "
-         "AND BaseFiles.FileId = File.FileId "
-         "AND Path.PathId=File.PathId",
-         edit_int64(jobid, ed1));
-  }
+  Mmsg(cmd,
+       "SELECT Path.Path||File.Name AS Filename "
+       "FROM BaseFiles, File, Path "
+       "WHERE BaseFiles.JobId=%s AND BaseFiles.BaseJobId = File.JobId "
+       "AND BaseFiles.FileId = File.FileId "
+       "AND Path.PathId=File.PathId",
+       edit_int64(jobid, ed1));
 
   sendit->ArrayStart("files");
   if (!BigSqlQuery(cmd, ::ListResult, &lctx)) { return; }
@@ -788,5 +760,4 @@ void BareosDb::ListFilesets(JobControlRecord* jcr,
 
   SqlFreeResult();
 }
-#endif /* HAVE_SQLITE3 || HAVE_MYSQL || HAVE_POSTGRESQL || HAVE_INGRES || \
-          HAVE_DBI */
+#endif /* HAVE_POSTGRESQL */