+++ docs/manuals/source/conf.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 #   BAREOS - Backup Archiving REcovery Open Sourced
 #
-#   Copyright (C) 2018-2022 Bareos GmbH & Co. KG
+#   Copyright (C) 2018-2023 Bareos GmbH & Co. KG
 #
 #   This program is Free Software; you can redistribute it and/or
 #   modify it under the terms of version three of the GNU Affero General Public
@@ -74,7 +74,6 @@ rst_epilog = """
 .. |mariadb| replace:: MariaDB
 .. |mysql| replace:: MySQL
 .. |postgresql| replace:: PostgreSQL
-.. |sqlite| replace:: Sqlite
 .. |vmware| replace:: VMware
 .. |vsphere| replace:: VMware vSphere
 
@@ -351,19 +350,10 @@ linkcheck_rate_limit_timeout = 300.00
 
 # -- Options for manual page output ---------------------------------------
 
-man_bareos_dbcopy = "man/bareos-dbcopy"
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (
-        man_bareos_dbcopy,
-        "bareos-dbcopy",
-        "Copy the Bareos catalog database between catalog backends",
-        [author],
-        8,
-    )
-]
+man_pages = []
 
 
 # -- Options for Texinfo output -------------------------------------------